Get the Tabula Muris Senis FACS single-cell RNA-seq data

Access the FACS (Smart-Seq2) RNA-seq data from the Tabula Muris Senis consortium.

TabulaMurisSenisFACS(
  tissues = "All",
  processedCounts = FALSE,
  reducedDims = TRUE,
  infoOnly = FALSE
)

Arguments

tissues

A character vector with the tissues to retrieve objects for. A list of available tissues can be obtained using listTabulaMurisSenisTissues("FACS").

processedCounts

Logical scalar. If TRUE, include the processed counts in addition to the raw counts in the SingleCellExperiment object.

reducedDims

Logical scalar. If TRUE, include the PCA, tSNE and UMAP representations in the SingleCellExperiment object (the tSNE representation is not available for the full dataset ('All' tissue)).

infoOnly

Logical scalar. If TRUE, only print the total size of the files that will be downloaded to and/or retrieved from the cache.

Value

If infoOnly is FALSE, returns a named list of

SingleCellExperiment objects (one per tissue requested via tissues). Otherwise, each element in the list is `NULL`.

Details

The data set was downloaded from figshare (https://figshare.com/articles/dataset/Processed_files_to_use_with_scanpy_/8273102?file=23937842 for the full data set, https://figshare.com/articles/dataset/Tabula_Muris_Senis_Data_Objects/12654728 for the individual tissue ones).

References

Schaum et al (2019): The murine transcriptome reveals global aging nodes with organ-specific phase and amplitude. bioRxiv doi:10.1101/662254.

The Tabula Muris Consortium (2020): A single-cell transcriptomic atlas characterizes ageing tissues in the mouse. Nature 583:590–595.

Author

Charlotte Soneson

Examples

if (interactive()) {
  sce <- TabulaMurisSenisFACS(tissues = "All")
}